+++ argopy/data_fetchers/proto.py
@@ -58,12 +58,16 @@ class ArgoDataFetcherProto(ABC):
         elif self.dataset_id in ['bgc', 'bgc-s']:
 
             if self.user_mode == "standard":
-                to_remove = []
+                to_remove = sorted(
+                    list(set(list(ds.data_vars)) - set(list_standard_variables(ds=self.dataset_id)))
+                )
                 [to_remove.append(v) for v in ds.data_vars if 'CDOM' in v]
                 return ds.drop_vars(to_remove)
 
             elif self.user_mode == "research":
-                to_remove = []
+                to_remove = sorted(
+                    list(set(list(ds.data_vars)) - set(list_standard_variables(ds=self.dataset_id)))
+                )
                 [to_remove.append(v) for v in ds.data_vars if 'CDOM' in v]
                 [to_remove.append(v) for v in ds.data_vars if "DATA_MODE" in v]
                 [to_remove.append(v) for v in ds.data_vars if "QC" in v]