+++ argopy/xarray.py
@@ -654,7 +654,7 @@ class ArgoAccessor:
         Parameters
         ----------
         params: str, List[str], optional, default='all'
-            Name or list of names of the parameter to merge.
+            Name or list of names of the parameter(s) to merge.
             Use the default keyword ``all`` to merge all possible parameters in the :class:`xarray.Dataset`.
         errors: str, optional, default='raise'
             If ``raise``, raises a :class:`argopy.errors.InvalidDatasetStructure` error if any of the expected variables is
@@ -682,37 +682,32 @@ class ArgoAccessor:
 
         # Determine the list of variables to transform:
         params = to_list(params)
+        parameters = []
+        # log.debug(params)
         if params[0] == "all":
             if "DATA_MODE" in this.data_vars:
-                params = ["PRES", "TEMP"]
-                if "PSAL" in this.data_vars:
-                    params.append("PSAL")
+                for p in list_core_parameters():
+                    if p in this.data_vars or "%s_ADJUSTED" % p in this.data_vars:
+                        parameters.append(p)
             else:
-                params = [
+                parameters = [
                     p.replace("_DATA_MODE", "")
                     for p in this.data_vars
                     if "_DATA_MODE" in p
                 ]
         else:
-            for p in params:
-                if p not in this.data_vars:
-                    if errors == "raise":
-                        raise InvalidDatasetStructure(
-                            "Parameter '%s' not found in this dataset" % p
-                        )
-                    else:
-                        log.debug("Parameter '%s' not found in this dataset" % p)
-                    params.remove(p)
+            [parameters.append(v) for v in params]
+        # log.debug(parameters)
 
         # Transform data:
-        for param in params:
+        for param in parameters:
             this = merge_param_with_param_adjusted(this, param, errors=errors)
 
         # Finalise:
         this = this[np.sort(this.data_vars)]
         this.argo.add_history(
             "[%s] real-time and adjusted/delayed variables merged according to their data mode"
-            % (",".join(params))
+            % (",".join(parameters))
         )
 
         return this
@@ -791,7 +786,7 @@ class ArgoAccessor:
                     if "_DATA_MODE" in p
                 ]
         elif params[0] == "core":
-            params = ["PRES", "TEMP", "PSAL"]
+            params = list_core_parameters()
         else:
             for p in params:
                 if p not in this.data_vars:
@@ -1087,7 +1082,7 @@ class ArgoAccessor:
             this = this.argo.profile2point()
 
         core_params = list_core_parameters()
-        if "PSAL" not in this.data_vars:
+        if "PSAL" not in this.data_vars and "PSAL_ADJUSTED" not in this.data_vars:
             core_params.remove("PSAL")
 
         # Apply transforms and filters: